chore: restructure code - dashboard views - roles
